scripts/requirements.txt
pandas
openpyxl
distance
Pillow
pdf2image
numpy